explore mode pickup and place
